Split inference code into multiple files
